# backend/swagger.yaml
openapi: 3.0.0
info:
  title: XWZ LTD Parking Management System API
  version: 1.0.0
  description: API documentation for the parking management system
servers:
  - url: http://localhost:5000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CarEntry:
      type: object
      required:
        - plateNumber
        - parking
        - entryTime
      properties:
        plateNumber:
          type: string
          description: Vehicle's plate number
        parking:
          type: string
          description: Reference to parking lot
        entryTime:
          type: string
          format: date-time
          description: Entry time
        exitTime:
          type: string
          format: date-time
          nullable: true
          description: Exit time (null if car is still parked)
        status:
          type: string
          enum: [active, completed]
          description: Entry status
        fee:
          type: number
          minimum: 0
          description: Charged amount in RWF
paths:
...
  /report/outgoing:
    get:
      tags:
        - Reports
      summary: Get outgoing cars report
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Outgoing cars report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CarEntry'
                  totalAmount:
                    type: number
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      pages:
                        type: integer
        '400':
          description: Invalid date range
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /report/entered:
    get:
      tags:
        - Reports
      summary: Get entered cars report
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Entered cars report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CarEntry'
                  statistics:
                    type: object
                    properties:
                      totalEntries:
                        type: integer
                      currentlyParked:
                        type: integer
                      totalRevenue:
                        type: number
                      averageDuration:
                        type: number
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      pages:
                        type: integer
        '400':
          description: Invalid date range
        '401':
          description: Unauthorized
        '500':
          description: Server error